Major fixes

+++ se/kth/iv1350/integration/SaleDTO.java
@@ -4,27 +4,14 @@ import java.time.LocalDate;
 import java.util.ArrayList;
 import se.kth.iv1350.model.Item;
 
-/**
- * Represents the details of a sale, including VAT, items, total amount, etc.
- */
 public class SaleDTO {
-
-    private int VAT;  // Total VAT for the sale
+    private double VAT;
     private ArrayList<Item> items;
     private LocalDate saleDate;
-    private int totalAmount;
-    private int amountPaid;
+    private double totalAmount;
+    private double amountPaid;
 
-    /**
-     * Constructor for SaleDTO.
-     * 
-     * @param VAT The total VAT for the sale.
-     * @param items The list of items in the sale.
-     * @param saleDate The date of the sale.
-     * @param amountPaid The amount paid by the customer.
-     * @param totalAmount The total price of the sale.
-     */
-    public SaleDTO(int VAT, ArrayList<Item> items, LocalDate saleDate, int amountPaid, int totalAmount) {
+    public SaleDTO(double VAT, ArrayList<Item> items, LocalDate saleDate, double amountPaid, double totalAmount) {
         this.VAT = VAT;
         this.items = items;
         this.saleDate = saleDate;
@@ -32,8 +19,8 @@ public class SaleDTO {
         this.totalAmount = totalAmount;
     }
 
-    public int getVAT() {
-        return VAT;  // Return the total VAT for the sale
+    public double getVAT() {
+        return VAT;
     }
 
     public ArrayList<Item> getItems() {
@@ -44,11 +31,11 @@ public class SaleDTO {
         return saleDate;
     }
 
-    public int getTotalAmount() {
+    public double getTotalAmount() {
         return totalAmount;
     }
 
-    public int getAmountPaid() {
+    public double getAmountPaid() {
         return amountPaid;
     }
 }
